モジュール構成の見直し

+++ ZipUtility/ZipEntryGeneralPurposeBitFlag.cs
@@ -7,13 +7,13 @@ namespace ZipUtility
         : UInt16
     {
         None = 0,
-        Encrypted = (UInt16)1 << 0,
-        CompresssionOption0 = (UInt16)1 << 1,
-        CompresssionOption1 = (UInt16)1 << 2,
-        HasDataDescriptor = (UInt16)1 << 3,
-        CompressedPatchedData = (UInt16)1 << 5,
-        StrongEncrypted = (UInt16)1 << 6,
-        UseUnicodeEncodingForNameAndComment = (UInt16)1 << 11,
-        EncryptedCentralDirectory = (UInt16)1 << 13
+        Encrypted = 1 << 0,
+        CompresssionOption0 = 1 << 1,
+        CompresssionOption1 = 1 << 2,
+        HasDataDescriptor = 1 << 3,
+        CompressedPatchedData = 1 << 5,
+        StrongEncrypted = 1 << 6,
+        UseUnicodeEncodingForNameAndComment = 1 << 11,
+        EncryptedCentralDirectory = 1 << 13
     }
 }
